blog/search_indexes.py
#coding=utf-8
__author__ = 'CoderSong'
__date__ = '2019/4/19 19:22'

from haystack import indexes
from .models import Article


class ArticleIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Article

    def index_queryset(self, using=None):
        return self.get_model().objects.all()